(fix): refactor main.go
